feat: add --skip-submodules argument
One could argue that this should even be the default because the
submodule is managed by containing repository. If the submodule is
dirty, the containing repository is dirty as well. Plus the submodule
might very well not be on a branch like master but on a specific git
commit and that is fine.

But better not make it the default as it would change the behavior. When
submodules annoy clustergit users, they can now use the
--skip-submodules argument